+++ src/Exception/PreviewTokenExpired.php
@@ -11,6 +11,8 @@ use function strpos;
 
 final class PreviewTokenExpired extends RequestFailure
 {
+    public const EXPECTED_ERROR_MESSAGE = 'Preview token expired';
+
     public static function isPreviewTokenExpiry(ResponseInterface $response) : bool
     {
         $type = $response->getHeaderLine('content-type');
@@ -21,15 +23,14 @@ final class PreviewTokenExpired extends RequestFailure
         $payload = Json::decodeObject((string) $response->getBody());
         $error = $payload->error ?? null;
 
-        return $error === 'Preview token expired';
+        return $error === self::EXPECTED_ERROR_MESSAGE;
     }
 
     public static function with(RequestInterface $request, ResponseInterface $response) : self
     {
-        $status = $response->getStatusCode();
         $error = new static(sprintf(
             'Error %d. The preview token provided has expired',
-            $status
+            $response->getStatusCode()
         ), $response->getStatusCode());
         $error->request = $request;
         $error->response = $response;